pyproject.toml
[project]
name = "arena-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.41.1,<2.0.0)",
    "anthropic (>=0.45.2,<0.46.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
